refactor: rename `snapshot_uploader` module to `file_uploader`
