+++ mithril-aggregator/src/file_uploaders/mod.rs
@@ -1,13 +1,13 @@
 mod dumb_snapshot_uploader;
+mod file_uploader;
 mod local_snapshot_uploader;
 mod remote_snapshot_uploader;
-mod snapshot_uploader;
 
 pub use dumb_snapshot_uploader::*;
+pub use file_uploader::SnapshotLocation;
+pub use file_uploader::SnapshotUploader;
 pub use local_snapshot_uploader::LocalSnapshotUploader;
 pub use remote_snapshot_uploader::RemoteSnapshotUploader;
-pub use snapshot_uploader::SnapshotLocation;
-pub use snapshot_uploader::SnapshotUploader;
 
 #[cfg(test)]
-pub use snapshot_uploader::MockSnapshotUploader;
+pub use file_uploader::MockSnapshotUploader;
